refact: move implementations to its .c file
