+++ tests/CMakeLists.txt
@@ -1,10 +1,10 @@
 add_executable(tests
         unity/src/unity.c   # Unity framework
         test_mdarray.c
+        ${CMAKE_SOURCE_DIR}/src/mdarray.c
 )
 
 # Include Unity headers
-target_include_directories(tests PRIVATE unity/src)
 target_include_directories(tests
         PRIVATE
         ${CMAKE_SOURCE_DIR}/src
